Party manager
The way we currently get and create new settlement info is dependant on their getters. Instead, we should do something similar to what PartyManager now does and watch the difference between our list count, and the game's list count. If they are different find out why, and deal with it.