Daily Behaviour and optimisations
I can't currently find a delegate "onPartySpawn" or something like that. I think that looping through all partyInfos every tick (assuming they're different in count to the party list) can impact performance. I managed to reduce it by always destroying partyinfo whenever a party is destroyed by using a delegate, but can't do the same for creating (hence the loop stays in place).

I've added DailyUpdate functions for settlement infos. Perhaps we should add them as virtual functions for the base classes?

I extended the SettlementInfoExtensions with GetGarrison and GetMilitia functions.

It's becoming clear we're going to need to create our own base mod campaign behaviour to utilise various variables. For example "ClanCanJoinOtherKingdom" should be called with a delegate, but right now it's up to the mod to implement this functionality. We need a base mod that will do all of this for us (same goes for all "WatchX" functions.